more debugging fixes

+++ include/Utility.h
@@ -363,7 +363,7 @@ inline const RE::BSFixedString findPriorityMatch(const RE::BSFixedString& nodeNa
 }
 
 //gets users skyrim pref setting. lights reinitialized are set to this exact distance
-inline void getObjectFadeMult(float& fLODFadeOutMultObjects) {
+inline void getObjectFadeMult(int& fLODFadeOutMultObjects) {
 
 	if (auto* setting = RE::GetINISetting("fLODFadeOutMultObjects:LOD")) {
 		if (setting->GetType() == RE::Setting::Type::kFloat) {
